+++ pkg/generator/dsl_convert.go
@@ -63,8 +63,8 @@ func dslTemplateToInput(tmpl string) (*models.CreateTransactionInput, error) {
 		}
 
 		to = append(to, models.FromToInput{
-			Account: d[2],
-			Share:   &models.Share{Percentage: pct},
+			AccountAlias: d[2],
+			Share:        &models.Share{Percentage: pct},
 		})
 	}
 
@@ -76,7 +76,7 @@ func dslTemplateToInput(tmpl string) (*models.CreateTransactionInput, error) {
 			Value: amount,
 			Source: &models.SourceInput{
 				From: []models.FromToInput{
-					{Account: src[1], Amount: models.AmountInput{Asset: asset, Value: amount}},
+					{AccountAlias: src[1], Amount: models.AmountInput{Asset: asset, Value: amount}},
 				},
 			},
 			Distribute: &models.DistributeInput{To: to},